# backend/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "NEXTAN Costing & Quote Generator API",
    "description": "Internal costing sheets and client-facing branded quote exports.",
    "version": "1.0.0"
  },
  "paths": {
...
    "/api/v1/scenarios/{scenario_id}/exports": {
      "post": {
        "tags": [
          "Exports"
        ],
        "summary": "Create Export",
        "description": "Purpose: Generate a quote export (DOCX or PDF). Steps:\n         1. Compute snapshot_data and render file (outside DB transaction)\n         2. Upload to Supabase Storage (outside DB transaction)\n         3. INSERT QuoteExport record inside DB transaction (revision_number computed inside txn)\nInputs: scenario_id (str UUID), ExportCreate (file_type)\nOutputs: ExportRead\nOwner: [Claude]",
        "operationId": "create_export_api_v1_scenarios__scenario_id__exports_post",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "scenario_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Scenario Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ExportCreate"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ExportRead"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "get": {
        "tags": [
          "Exports"
        ],
        "summary": "List Scenario Exports",
        "description": "Purpose: List exports for a specific scenario.\nInputs: scenario_id (str UUID)\nOutputs: list[ExportRead]\nOwner: [Claude]",
        "operationId": "list_scenario_exports_api_v1_scenarios__scenario_id__exports_get",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "scenario_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Scenario Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/ExportRead"
                  },
                  "title": "Response List Scenario Exports Api V1 Scenarios  Scenario Id  Exports Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/costing-sheets/{sheet_id}/exports": {
      "get": {
        "tags": [
          "Exports"
        ],
        "summary": "List Sheet Exports",
        "description": "Purpose: List all exports for a costing sheet across all scenarios and file types.\nInputs: sheet_id (str UUID)\nOutputs: list[ExportRead]\nOwner: [Claude]",
        "operationId": "list_sheet_exports_api_v1_costing_sheets__sheet_id__exports_get",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "sheet_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Sheet Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/ExportRead"
                  },
                  "title": "Response List Sheet Exports Api V1 Costing Sheets  Sheet Id  Exports Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/exports/{export_id}/download": {
      "get": {
        "tags": [
          "Exports"
        ],
        "summary": "Download Export",
        "description": "Purpose: Return a signed Supabase Storage URL for downloading an export file.\nInputs: export_id (str UUID)\nOutputs: ExportDownloadResponse (signed_url, expires_in_seconds, metadata)\nOwner: [Claude]",
        "operationId": "download_export_api_v1_exports__export_id__download_get",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "export_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Export Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ExportDownloadResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/exports/{export_id}": {
      "delete": {
        "tags": [
          "Exports"
        ],
        "summary": "Delete Export",
        "description": "Purpose: Delete an export record and its Supabase Storage file.\n         After deletion, sheet/scenario may become deletable if no exports remain.\nInputs: export_id (str UUID)\nOutputs: 204 No Content\nOwner: [Claude]",
        "operationId": "delete_export_api_v1_exports__export_id__delete",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "export_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Export Id"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Successful Response"
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "ExportCreate": {
        "properties": {
          "file_type": {
            "type": "string",
            "title": "File Type"
          }
        },
        "type": "object",
        "required": [
          "file_type"
        ],
        "title": "ExportCreate",
        "description": "Purpose: Body for POST /scenarios/{scenario_id}/exports.\nInputs: file_type ('docx' | 'pdf')\nOutputs: N/A\nOwner: [Claude]"
      },
      "ExportDownloadResponse": {
        "properties": {
          "export_id": {
            "type": "string",
            "format": "uuid",
            "title": "Export Id"
          },
          "signed_url": {
            "type": "string",
            "title": "Signed Url"
          },
          "expires_in_seconds": {
            "type": "integer",
            "title": "Expires In Seconds"
          },
          "file_type": {
            "type": "string",
            "title": "File Type"
          },
          "revision_number": {
            "type": "integer",
            "title": "Revision Number"
          }
        },
        "type": "object",
        "required": [
          "export_id",
          "signed_url",
          "expires_in_seconds",
          "file_type",
          "revision_number"
        ],
        "title": "ExportDownloadResponse",
        "description": "Purpose: Response for GET /exports/{export_id}/download.\n         Returns a signed Supabase Storage URL valid for a short TTL.\nInputs: N/A (response body)\nOutputs: signed_url, expires_in_seconds, export metadata\nOwner: [Claude]"
      },
      "ExportRead": {
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid",
            "title": "Id"
          },
          "costing_sheet_id": {
            "type": "string",
            "format": "uuid",
            "title": "Costing Sheet Id"
          },
          "scenario_id": {
            "type": "string",
            "format": "uuid",
            "title": "Scenario Id"
          },
          "user_id": {
            "type": "string",
            "format": "uuid",
            "title": "User Id"
          },
          "revision_number": {
            "type": "integer",
            "title": "Revision Number"
          },
          "file_type": {
            "type": "string",
            "title": "File Type"
          },
          "file_path": {
            "type": "string",
            "title": "File Path"
          },
          "exported_at": {
            "type": "string",
            "format": "date-time",
            "title": "Exported At"
          }
        },
        "type": "object",
        "required": [
          "id",
          "costing_sheet_id",
          "scenario_id",
          "user_id",
          "revision_number",
          "file_type",
          "file_path",
          "exported_at"
        ],
        "title": "ExportRead",
        "description": "Purpose: QuoteExport record in list responses.\n         Does not include snapshot_data to keep payload small.\nInputs: N/A (response body)\nOutputs: id, costing_sheet_id, scenario_id, user_id, revision_number, file_type, file_path, exported_at\nOwner: [Claude]"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          },
          "input": {
            "title": "Input"
          },
          "ctx": {
            "type": "object",
            "title": "Context"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      }
    },
    "securitySchemes": {
      "HTTPBearer": {
        "type": "http",
        "scheme": "bearer"
      }
    }
  }
}